requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

psycopg2-binary==2.9.9

pytest==8.1.1
pytest-bdd==7.1.2
requests==2.31.0
